+++ server/prisma/migrations/20240812133954_keep_track_of_clan_war_wins/migration.sql
@@ -1,0 +1,3 @@
+-- AlterTable
+ALTER TABLE "ClanWar" ADD COLUMN     "attackerWins" INTEGER NOT NULL DEFAULT 0,
+ADD COLUMN     "defenderWins" INTEGER NOT NULL DEFAULT 0;